# dc3-common/dc3-common-auth/src/main/resources/openapi/openapi-agentic.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "IoT DC3 REST API",
        "description": "IoT DC3 platform REST API documentation.\n\nCovers all service modules: auth (authentication and authorization),\nmanager (device/driver/point/profile configuration), data (telemetry,\nevents, commands, notifications), and agentic (AI chat and tools).\n\nAuthentication: obtain a salt and token via /api/v3/auth/token/*, then\nuse the Authorize button to set X-Auth-Tenant, X-Auth-Login, and\nX-Auth-Token. X-Auth-Token is a JSON object {\"salt\":\"...\",\"token\":\"...\"}.\n",
        "contact": {
            "name": "pnoker",
            "url": "https://github.com/pnoker/iot-dc3",
            "email": "pnokers@icloud.com"
        },
        "license": {
            "name": "GNU Affero General Public License v3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.txt"
        },
        "version": "2026.6.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:18600/agentic",
            "description": "Generated server url"
        }
    ],
    "security": [
        {
            "X-Auth-Tenant": [],
            "X-Auth-Login": [],
            "X-Auth-Token": []
        }
    ],
...
    "paths": {
...
        "/provider/config/update": {
            "post": {
                "tags": [
                    "provider"
                ],
                "summary": "Update Model Provider",
                "description": "Update an existing LLM provider for the current tenant, changing its base URL, credentials or capability spec. Returns the updated provider; the target must belong to the current tenant.",
                "operationId": "update_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModelProviderVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelProviderVO"
                                }
                            }
                        }
                    }
                },
                "x-dc3-ai": {
                    "idempotent": "true",
                    "openWorld": "false",
                    "riskLevel": "MEDIUM",
                    "destructive": "false"
                }
            }
        },
        "/provider/config/add": {
            "post": {
                "tags": [
                    "provider"
                ],
                "summary": "Add Model Provider",
                "description": "Register a new upstream LLM provider for the current tenant with its base URL, credentials and capability spec. Returns the created provider; reference it afterwards when defining model configurations.",
                "operationId": "add",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModelProviderVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelProviderVO"
                                }
                            }
                        }
                    }
                },
                "x-dc3-ai": {
                    "idempotent": "false",
                    "openWorld": "false",
                    "riskLevel": "MEDIUM",
                    "destructive": "false"
                }
            }
        },
        "/model/config/update": {
            "post": {
                "tags": [
                    "model"
                ],
                "summary": "Update AI Model Configuration",
                "description": "Update an existing AI model configuration's provider, model id or parameters for the current tenant. Returns the updated configuration; changes apply to future sessions that select this model.",
                "operationId": "update_2",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModelConfigVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelConfigVO"
                                }
                            }
                        }
                    }
                },
                "x-dc3-ai": {
                    "idempotent": "true",
                    "openWorld": "false",
                    "riskLevel": "MEDIUM",
                    "destructive": "false"
                }
            }
        },
        "/model/config/add": {
            "post": {
                "tags": [
                    "model"
                ],
                "summary": "Add AI Model Configuration",
                "description": "Create an AI model configuration for the current tenant binding a provider, model id and parameters. Returns the saved configuration; the assistant can then use it as a selectable model in a session.",
                "operationId": "add_1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ModelConfigVO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ModelConfigVO"
                                }
                            }
                        }
                    }
                },
                "x-dc3-ai": {
                    "idempotent": "false",
                    "openWorld": "false",
                    "riskLevel": "MEDIUM",
                    "destructive": "false"
                }
            }
        },
...
        "/model/config/list": {
            "get": {
                "tags": [
                    "model"
                ],
                "summary": "List AI Model Configurations",
                "description": "List the stored AI model configurations for the current tenant, each binding a provider, model id and parameters. Returns full configuration records for management, unlike the lightweight model options used for chat selection.",
                "operationId": "listConfigs",
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ModelConfigVO"
                                    }
                                }
                            }
                        }
                    }
                },
                "x-dc3-ai": {
                    "idempotent": "true",
                    "openWorld": "false",
                    "riskLevel": "LOW",
                    "destructive": "false"
                }
            }
        },
...
        "/provider/config/delete": {
            "delete": {
                "tags": [
                    "provider"
                ],
                "summary": "Delete Model Provider",
                "description": "Permanently remove an LLM provider from the current tenant by id. Returns true on success; model configurations bound to this provider will no longer resolve, so call only when the provider is unused.",
                "operationId": "delete_1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Primary key of the entity to delete. Must belong to the current tenant.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1024
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                },
                "x-dc3-ai": {
                    "idempotent": "true",
                    "openWorld": "false",
                    "riskLevel": "HIGH",
                    "destructive": "true"
                }
            }
        },
        "/model/config/delete": {
            "delete": {
                "tags": [
                    "model"
                ],
                "summary": "Delete AI Model Configuration",
                "description": "Permanently delete the AI model configuration identified by id within the current tenant. Returns true on success; the model is no longer selectable in new sessions.",
                "operationId": "delete_2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "Primary key of the entity to delete. Must belong to the current tenant.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": 1024
                    }
                ],
                "responses": {
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProblemDetailsResponse"
                                }
                            }
                        }
                    },
                    "204": {
                        "description": "No Content"
                    }
                },
                "x-dc3-ai": {
                    "idempotent": "true",
                    "openWorld": "false",
                    "riskLevel": "HIGH",
                    "destructive": "true"
                }
            }
        }
    },
    "components": {
        "schemas": {
            "ProblemDetailsResponse": {
                "type": "object",
                "properties": {
                    "type": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "status": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "code": {
                        "type": "string"
                    },
                    "detail": {
                        "type": "string"
                    },
                    "instance": {
                        "type": "string"
                    },
                    "traceId": {
                        "type": "string"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "X-Auth-Tenant": {
                "type": "apiKey",
                "description": "Tenant code, e.g. 'default'.",
                "name": "X-Auth-Tenant",
                "in": "header"
            },
            "X-Auth-Login": {
                "type": "apiKey",
                "description": "Login name, e.g. 'dc3'.",
                "name": "X-Auth-Login",
                "in": "header"
            },
            "X-Auth-Token": {
                "type": "apiKey",
                "description": "Token envelope JSON: {\"salt\":\"...\",\"token\":\"...\"}.",
                "name": "X-Auth-Token",
                "in": "header"
            }
        }
    }
}